requirements.txt
requests~=2.28.1
pyaudio~=0.2.12
azure-cognitiveservices-speech
numpy
google-cloud-speech
gTTS
playsound
python-dotenv
git+https://github.com/lucregrassi/cairlib.git#egg=cairlib
